backend/requirements.txt
# Django & Base
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
requests
gunicorn
whitenoise

# Google AI (Gemini & LangChain)
google-generativeai
langchain
langchain-community
langchain-google-genai

# Image Processing (For OCR)
Pillow

# RAG & Vectores (FAISS + HuggingFace)
faiss-cpu
langchain-huggingface
sentence-transformers
chromadb

# calculus library for CrossEncoder & Embeddings
torch
numpy

# HuggingFace & LangChain
langchain
langchain-community
huggingface-hub